--- src/ducks/Signals/chart/newSignalLine/SignalLine.module.scss
@@ -1,31 +1,0 @@
-@import '~@santiment-network/ui/mixins';
-@import '~@santiment-network/ui/variables';
-
-.line {
-  border: 1px dashed var(--rhino);
-}
-
-.tooltip {
-  padding: 2px 8px;
-  background: var(--rhino);
-  color: var(--casper);
-  position: relative;
-  left: 8px;
-  bottom: 5px;
-  border-radius: $border-radius;
-  display: table;
-
-  @include text('caption');
-}
-
-.container {
-  pointer-events: none;
-  width: calc(100% - 40px);
-  position: absolute;
-  transform: translateX(40px) translateY(calc(var(--y) - 23px));
-  z-index: 1;
-}
-
-.highline {
-  color: var(--white);
-}
